Rename to Hyrule Builder++ and modify project layout to reflect changes
