+++ hyrule_builder_plus/data/switch/rstb.json
